fix: use long instead of int for ids for better spring integration

+++ src/main/java/org/spine/iquestionapi/model/Questionnaire.java
@@ -22,7 +22,7 @@ import lombok.Setter;
 public class Questionnaire implements Model {
     @Id
     @GeneratedValue
-    private int id;
+    private long id;
     private String name;
     @OneToMany(cascade = CascadeType.ALL)
     private List<Segment> segments = new ArrayList<>();
